Add float/double rating support with comma and dot decimal separators
🔢 NEW FEATURE: Float ratings with flexible input formats

Supported formats (as requested):
✅ 2.0   → 2.0
✅ 2     → 2.0
✅ 2.3   → 2.3
✅ 2,3   → 2.3 (comma replaced with dot)
✅ 0.7   → 0.7
✅ 1,56  → 1.56

Technical changes:
- Database: rating INTEGER → FLOAT
- Added parse_rating() function with comma/dot support
- Added is_valid_rating() validation
- Updated all setrating command variants
- Auto-rounding to 2 decimal places
- Support for negative ratings

Updated commands:
- /setrating 2,5 (self)
- /setrating @username 2,3 (admin)
- /setrating 123456789 1.75 (admin)
- /createuser 123456789 2,5 player (admin)

Testing:
✅ 10 new tests for float functionality
✅ All required formats tested and working
✅ Database schema updated correctly
✅ Backward compatibility maintained

European (comma) and American (dot) decimal formats both supported!
Perfect for international users.

+++ app/models/database.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
 from sqlalchemy.orm import sessionmaker
@@ -16,7 +16,7 @@ class UserRating(Base):
     telegram_username = Column(String(255), nullable=True, index=True, comment="Telegram @username")
     first_name = Column(String(255), nullable=True, comment="Telegram first name")
     PT_userId = Column(String(255), nullable=True, comment="PlayTomic username")
-    rating = Column(Integer, default=0)
+    rating = Column(Float, default=0.0)
     created_at = Column(DateTime, default=datetime.utcnow)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
